add contrail-docker-push-base job
This is intended as a new parent instead of c-z-j docker-base.
The job is placed in c-p-c to support secret for pushing
to an authenticated registry; now we can only push anonymously.

The pre-run is copied from c-z-j docker-base verbatim; it worked there.
The post-run is copied from contrail-build-containers-ci-base,
where it also worked.

The contrail-build-containers-ci-base will be useless as it's placed
too low in the hierarchy and unusable by contrail-build-test-containers.

Change-Id: I1edfb2bab9bceb90056fbe220f3fe858893df18a