build numbering sequence per pipeline
We cannot use new `periodic-nightly-test` pipeline yet, because
it would use the same sequence of numbering as `periodic-nightly`.
It could confuse users.

The `version` in fact determines the sequence to use and up to now it
was a branch, either 'master' or 'R5.0'.
Before: we had a numbering sequence per branch.
After: we'll have a numbering sequence per branch+pipeline.

To ensure stability, this change impacts only
pipeline `periodic-nightly-test` (which will start builds from number 1)
but it leaves the legacy behavior for all other pipelines (no change
in numbering).

Partial-Jira-bug: JD-428

Change-Id: Ic23e68e755cf1def6aba7b96b6e58b2185ba786e