bugfix port for nightly Nexus registry
The periodic-nightly and periodic-nightly-test tried to push to 5010,
which was never possible. Correcting port to 5002.

This change doesn't actually impact the production periodic-nightly
yet, because it doesn't run the modified role. This change is a blocker
to test migration of nightly builds to Nexus docker registry.

Debt: Old code was very cryptic with extra 3 (5) files and 1 directory.
Merely five lines of readable code sufficed to replace these.

Debt: Single quotes per style guide.

Change-Id: Ica400908a122a1f1fd9594811644fd864dbbd264
Partial-Jira-Bug: JD-428